+++ core/instrument/test/fixtures/csf/exclude-stories/include-stories-regex.js
@@ -1,8 +1,0 @@
-export default {
-  title: 'Storybook/Blocks/IncludeStoryRegex',
-  includeStories: /story(2|3)/,
-};
-
-export const story1 = () => {};
-export const story2 = () => {};
-export const story3 = () => {};