+++ integrations/gatsby-theme-stories/src/gatsby-node.ts
@@ -23,7 +23,6 @@ exports.createPages = async (
     webPack: options.webpack,
     presets: defaultPresets,
     configPath: options.configPath,
-    // bundleAnalyzer: true,
   };
   const pageTemplates: Record<PageType, string> = {
     story: require.resolve(`../src/templates/DocPage.tsx`),