+++ ui/blocks/src/PropsTable/controlsActions.ts
@@ -40,21 +40,21 @@ export const useControlsActions = (props: UseControlsActionsProps) => {
   };
   return [
     {
-      title: copied ? 'values copied' : 'copy values',
+      node: copied ? 'values copied' : 'copy values',
       onClick: onCopy,
       group: 'controls',
       id: 'copy_controls',
       'aria-label': 'copy control values',
     },
     {
-      title: 'reset',
+      node: 'reset',
       onClick: onReset,
       group: 'controls',
       id: 'reset',
       'aria-label': 'reset control values to their initial value',
     },
     {
-      title: 'randomize',
+      node: 'randomize',
       group: 'controls',
       onClick: () => {
         if (setControlValue && controls && storyId) {